fast laziness is separated
