+++ functional/settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'function'
+rootProject.name = 'functional'
