+++ functional/src/main/java/com/xosmig/fastfunctional/FastLazyList.java
@@ -1,0 +1,8 @@
+package com.xosmig.fastfunctional;
+
+import java.util.*;
+
+public interface FastLazyList<R> extends List<R> {
+    void takeAll();
+    void takeTo(int index);
+}
