--- functional/src/main/java/com/xosmig/function/Lazy.java
@@ -1,10 +1,0 @@
-package com.xosmig.function;
-
-public interface Lazy<T> {
-    T get();
-
-    static <T>
-    Lazy<T> value(T x) {
-        return () -> x;
-    }
-}
