+++ functional/src/main/java/com/xosmig/functional/Lazy.java
@@ -1,0 +1,5 @@
+package com.xosmig.functional;
+
+public interface Lazy<T> {
+    T value();
+}
