+++ functional/src/main/java/com/xosmig/functional/LazyValue.java
@@ -1,0 +1,14 @@
+package com.xosmig.functional;
+
+public class LazyValue<T> implements Lazy<T> {
+    T value;
+
+    public LazyValue(T value) {
+        this.value = value;
+    }
+
+    @Override
+    public T value() {
+        return value;
+    }
+}
