+++ functional/src/test/java/com/xosmig/functional/Function2Test.java
@@ -1,5 +1,7 @@
-package com.xosmig.function;
+package com.xosmig.functional;
 
+import com.xosmig.function.Function1;
+import com.xosmig.function.Function2;
 import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.is;
