+++ functional/src/test/java/com/xosmig/functional/Predicate1Test.java
@@ -1,5 +1,6 @@
-package com.xosmig.function;
+package com.xosmig.functional;
 
+import com.xosmig.function.Predicate1;
 import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.is;
